requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
httpx>=0.27.0
